Add initial prototype with hardcoded choices
This is just a start, we've talked about some
alternatives. It seems like the template system is
going to be critical to have sooner rather than
later, so the hardcoded bits will hopefully be
a small part of the final version.

This also kicks off the Cargo crate, though we may
end up splitting the crate into pieces later.