Add a dynamic plaintext table generator
The idea is you write a series of functions for getting all manner of properties
off some struct and render them into columns. This is a dependency of the code
for the character stats; it's used to print stat blocks

The tests end up being fairly messy given whitespace issues,
so rather than writing additional ones here, I'm thinking we
could rely on the ones in the stat block code (they
are written).