Improve stat progression curves and utilities
There's now a method for performing a check and
gaining progression from doing so, with the progression
scaled based on a target difficulty as discussed.

I also switched progress to an exponential curve
that is sublinear for most of the levels characters
may have but gives a soft cap towards the end.

The new code has some unit tests and I put a struct
for the player in the world that has a stat block.