Add a skeleton of the story graph
The story graph will hold the structure of the story
and control the flow. It will be the output of the
template parser.

Ultimately, we'll want more advanced templates to
replace most of the strings here, but this should
capture the basic idea. More implementation and tests
for the execution of the story will come later.